+++ inference_perf/prompts/base.py
@@ -4,7 +4,7 @@ import aiohttp
 import numpy as np
 from pydantic import BaseModel
 from inference_perf.collectors.request_lifecycle import PromptLifecycleMetric
-from inference_perf.colletor_metrics.base import ResponseData
+from inference_perf.colletor_metric.base import ResponseData
 from inference_perf.utils.custom_tokenizer import CustomTokenizer
