+++ inference_perf/prompts_metrics/base.py
@@ -1,19 +1,8 @@
-from abc import abstractmethod
 from typing import Any, Optional
+
 from pydantic import BaseModel
-
+from inference_perf.colletor_metric.base import Metric
 from inference_perf.prompts.base import LlmPrompt
-
-
-class Metric(BaseModel):
-    """Abstract type to track reportable (but not neccesarily summarizable) metrics"""
-
-    stage_id: Optional[int] = None
-
-    @abstractmethod
-    async def to_report(self) -> dict[str, Any]:
-        """Create the report for this metric"""
-        raise NotImplementedError
 
 
 class FailedResponseData(BaseModel):
